requirements.txt
paho-mqtt==1.5.0
six==1.15.0
crython==0.2.0
requests==2.25.1